refactor(injector): split injector from filter
Make sure injector is executed before after_route_render
