+++ lib/hexo/index.js
@@ -25,9 +25,9 @@ const defaultConfig = require('./default_config');
 const loadDatabase = require('./load_database');
 const multiConfigPath = require('./multi_config_path');
 const { sync } = require('resolve');
-const full_url_for = require('../plugins/helper/full_url_for');
+const injectorFilter = require('../plugins/injector');
 const { inherits } = require('util');
-const { deepMerge } = require('hexo-util');
+const { deepMerge, full_url_for } = require('hexo-util');
 
 const libDir = dirname(__dirname);
 const dbVersion = 1;
@@ -57,6 +57,7 @@ const createLoadThemeRoute = function(generatorResult, locals, ctx) {
       if (view) {
         log.debug(`Rendering HTML ${name}: ${magenta(path)}`);
         return view.render(locals)
+          .then(result => injectorFilter.call(ctx, result, locals))
           .then(result => ctx.execFilter('after_route_render', result, {
             context: ctx,
             args: [locals]
