+++ lib/plugins/filter/after_render/index.js
@@ -5,5 +5,4 @@ module.exports = ctx => {
 
   filter.register('after_route_render', require('./external_link'));
   filter.register('after_route_render', require('./meta_generator'));
-  filter.register('after_route_render', require('./injector'));
 };
