+++ lib/plugins/injector/index.js
