+++ test/index.js
@@ -10,6 +10,7 @@ describe('Hexo', () => {
   require('./scripts/generators');
   require('./scripts/helpers');
   require('./scripts/hexo');
+  require('./scripts/injector');
   require('./scripts/models');
   require('./scripts/processors');
   require('./scripts/renderers');
