+++ test/scripts/filters/index.js
@@ -5,7 +5,6 @@ describe('Filters', () => {
   require('./excerpt');
   require('./external_link');
   require('./i18n_locals');
-  require('./injector');
   require('./meta_generator');
   require('./new_post_path');
   require('./post_permalink');
